requirement.txt
# Core Django framework
Django>=4.1,<5.0

# Django REST Framework for building APIs
djangorestframework>=3.14,<4.0

# CORS Headers to allow cross-origin requests from frontend
django-cors-headers>=3.13.0

# Python Dotenv for environment variable support
python-dotenv>=0.21.0

# MongoDB Database adapter
djongo>=1.3.6
pymongo==3.12.3

# For JSON Web Token Authentication (optional, if you implement JWT)
djangorestframework-simplejwt>=5.2.2

# Pillow to handle image fields (if uploading images)
Pillow>=9.3.0

# Gunicorn for production WSGI server (optional)
gunicorn>=20.1.0

# Optional: drf-yasg for API documentation
drf-yasg>=1.21.4

# Optional: Django Debug Toolbar for local dev debugging
django-debug-toolbar>=4.1

# Other dependencies you might add as needed
